+++ documentchannel-kafka/src/main/java/no/mnemonic/messaging/documentchannel/kafka/KafkaDocument.java
@@ -1,5 +1,9 @@
 package no.mnemonic.messaging.documentchannel.kafka;
 
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+
+import static no.mnemonic.commons.utilities.ObjectUtils.ifNotNull;
+
 /**
  * A transport object representing a document coming from Kafka
  *
@@ -7,16 +11,20 @@ package no.mnemonic.messaging.documentchannel.kafka;
  */
 public class KafkaDocument<T> {
 
-  private final T document;
+  private final ConsumerRecord<String, T> rec;
   private final String cursor;
 
-  public KafkaDocument(T document, String cursor) {
-    this.document = document;
+  public KafkaDocument(ConsumerRecord<String, T> rec, String cursor) {
+    this.rec = rec;
     this.cursor = cursor;
   }
 
   public T getDocument() {
-    return document;
+    return ifNotNull(rec, ConsumerRecord::value);
+  }
+
+  public ConsumerRecord<String, T> getRecord() {
+    return rec;
   }
 
   public String getCursor() {